+++ docs/changelog.rst
@@ -5,7 +5,9 @@ Changelog
 Changes in dev
 ==============
 
-- Added slave_okay kwarg to queryset
+- Added get_FIELD_display() method for easy choice field displaying.
+- Added queryset.slave_okay(enabled) method
+- Updated queryset.timeout(enabled) and queryset.snapshot(enabled) to be chainable
 - Added insert method for bulk inserts
 - Added blinker signal support
 - Added query_counter context manager for tests